[S07P12B102-313] BE / [커칵, 슈커칵] 커칵, 슈커칵 작성시 제목 미입력 검증 크로스 체크

+++ backend/src/main/java/com/jourgeois/backend/service/PostService.java
@@ -17,6 +17,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -55,27 +57,39 @@ public class PostService {
 
     @Transactional
     public void postPost(PostDTO postDTO) throws IOException, NoSuchElementException {
+
+
         Long writerId = postDTO.getUid();
 
         Member member = memberRepository.findById(writerId)
                             .orElseThrow(()->new NoSuchElementException("유저를 찾을 수 없습니다."));
-        // post 이미지 업로드
-        String uploadURL = s3Util.upload(postDTO.getImg(), member.getUid(), ImgType.POST);
         if(postDTO.getTitle()==null || postDTO.getTitle().isEmpty()){
+
             Post post = new Post();
             post.setDescription(postDTO.getDescription());
             post.setMember(member);
+            // post 이미지 업로드
+            String uploadURL = s3Util.upload(postDTO.getImg(), member.getUid(), ImgType.POST);
+
             post.setImg(uploadURL);
 
             // post 저장
             postRepository.save(post);
         }else{
+            // 제목, 커스텀 칵테일의 경우 재료, 레시피 공백 예외 처리
+            if(postDTO.getTitle().trim().isEmpty() || postDTO.getIngredients() == null || postDTO.getIngredients().length == 0 || postDTO.getRecipe().trim().isEmpty()){
+                throw  new IllegalArgumentException("필수 정보 누락");
+            }
+
             CustomCocktail cc = new CustomCocktail();
             cc.setTitle(postDTO.getTitle());
             cc.setDescription(postDTO.getDescription());
             cc.setMember(member);
             cc.setIngredients(Arrays.toString(postDTO.getIngredients()).replace("[","").replace("]",""));
             cc.setRecipe(postDTO.getRecipe());
+            // post 이미지 업로드
+            String uploadURL = s3Util.upload(postDTO.getImg(), member.getUid(), ImgType.POST);
+
             cc.setImg(uploadURL);
             // post 저장
             CustomCocktail cocktail = postRepository.save(cc);
@@ -101,8 +115,6 @@ public class PostService {
     @Transactional
     public void editPost(PostDTO postDTO) throws IOException, NoSuchElementException {
         Long postId = postDTO.getPostId();
-        // post 이미지 업로드
-        String uploadURL = s3Util.upload(postDTO.getImg(), postDTO.getUid(), ImgType.POST);
 
         Member member = new Member();
         member.setUid(postDTO.getUid());
@@ -112,13 +124,25 @@ public class PostService {
             Post targetPost = postRepository.findByIdAndMember(postId, member).orElseThrow(()->new NoSuchElementException("게시글을 찾을 수 없습니다."));
             targetPost.setDescription(postDTO.getDescription());
 
+            // post 이미지 업로드
+            String uploadURL = s3Util.upload(postDTO.getImg(), postDTO.getUid(), ImgType.POST);
+
             targetPost.setImg(uploadURL);
             s3Util.deleteFile(targetPost.getImg());
             // post 저장
             postRepository.save(targetPost);
         }else{
+            // 커스텀 칵테일의 경우 재료, 레시피 공백 예외 처리
+            if(postDTO.getTitle().trim().isEmpty() || postDTO.getIngredients() == null || postDTO.getIngredients().length == 0 || postDTO.getRecipe().trim().isEmpty()){
+                throw  new IllegalArgumentException("필수 정보 누락");
+            }
+
             CustomCocktail targetCustomCocktail = (CustomCocktail) postRepository.findByIdAndMember(postId, member).orElseThrow(()->new NoSuchElementException("게시글을 찾을 수 없습니다."));
             targetCustomCocktail.setDescription(postDTO.getDescription());
+
+            // post 이미지 업로드
+            String uploadURL = s3Util.upload(postDTO.getImg(), postDTO.getUid(), ImgType.POST);
+
             targetCustomCocktail.setImg(uploadURL);
             s3Util.deleteFile(targetCustomCocktail.getImg());
             targetCustomCocktail.setTitle(postDTO.getTitle());
